requirements.txt
python-dotenv
google-generativeai
streamlit
pdf2image
